requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
openai==0.27.0
python-nmap==0.7.1
requests==2.28.2
tqdm==4.65.0
urllib3==1.26.14
yarl==1.8.2
dnspython==2.3.0
rich==13.4.2
cowsay==5.0
tk==0.1.0
customtkinter==5.2.0
python-dotenv==1.0.0
transformers==4.31.0
auto-gptq==0.3.0
langchain==0.0.237
gradio==3.37.0
llama-cpp-python==0.1.73
fire==0.5.0
